% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_builder.R
\name{add_json_extension}
\alias{add_json_extension}
\title{Adds the json extension to the given url}
\usage{
add_json_extension(url)
}
\arguments{
\item{url}{base url to which the json extension should be added}
}
\value{
url with the json extension added
}
\description{
Adds the json extension to the given url
}
