% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_builder.R
\name{add_path}
\alias{add_path}
\title{Adds the given path to the given url}
\usage{
add_path(url, path)
}
\arguments{
\item{url}{base url to which the path should be added}

\item{path}{path that should be added to the url}
}
\value{
url with the given path added
}
\description{
Adds the given path to the given url
}
