% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion.R
\name{assert}
\alias{assert}
\title{Asserts a given expression and throws an error
if it returns FALSE}
\usage{
assert(expression, error)
}
\arguments{
\item{expression}{R expression to be evaluated}

\item{error}{message to be displayed when the
expression is not fulfilled}
}
\description{
Asserts a given expression and throws an error
if it returns FALSE
}
