% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion.R
\name{assert_ids}
\alias{assert_ids}
\title{Checks whether ids are correctly defined. If not throws an error}
\usage{
assert_ids(ids)
}
\arguments{
\item{ids}{maximum number of ids to retrieve}
}
\description{
Checks whether ids are correctly defined. If not throws an error
}
