% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion.R
\name{assert_max_items}
\alias{assert_max_items}
\title{Checks whether max_items is correctly defined. If not throws an error}
\usage{
assert_max_items(max_items)
}
\arguments{
\item{max_items}{maximum number of ids to retrieve}
}
\description{
Checks whether max_items is correctly defined. If not throws an error
}
