% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comment_to_dataframe_row}
\alias{comment_to_dataframe_row}
\title{Converts comment to a dataframe row}
\usage{
comment_to_dataframe_row(comment)
}
\arguments{
\item{comment}{comment to be converted to dataframe}
}
\value{
a dataframe containing a single entry with data from passed comment
}
\description{
Converts comment to a dataframe row
}
