% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hn_api_response.R
\name{create_hn_api_response}
\alias{create_hn_api_response}
\title{Creates an object representing a response from the
Hacker News API}
\usage{
create_hn_api_response(response)
}
\arguments{
\item{response}{response object}
}
\value{
object representing a response from the
Hacker News API
}
\description{
Creates an object representing a response from the
Hacker News API
}
