% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hn_item.R
\name{create_hn_item}
\alias{create_hn_item}
\title{Creates an object representing Hacker News item}
\usage{
create_hn_item(item)
}
\arguments{
\item{item}{list returned by Hacker News API}
}
\value{
object representing a hn_item
}
\description{
Creates an object representing Hacker News item
}
