% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hn_user.R
\name{create_hn_user}
\alias{create_hn_user}
\title{Creates an object representing Hacker News user}
\usage{
create_hn_user(user)
}
\arguments{
\item{user}{list returned by Hacker News API
via get_user_by_username function}
}
\value{
object representing a hn_user
}
\description{
Creates an object representing Hacker News user
}
