% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_builder.R
\name{create_request_url}
\alias{create_request_url}
\title{Creates a request url based on the given base url and
passed paths. The json extensions is added automatically.}
\usage{
create_request_url(url, paths)
}
\arguments{
\item{url}{base url of the request}

\item{paths}{vector of paths that should be added to the url}
}
\value{
request url with added paths and the json extension
}
\description{
Creates a request url based on the given base url and
passed paths. The json extensions is added automatically.
}
