% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_if_null}
\alias{default_if_null}
\title{return specified variable or default value if specified variable is null}
\usage{
default_if_null(variable, default)
}
\arguments{
\item{variable}{specified variable}

\item{default}{default return value}
}
\value{
specified variable or default value
}
\description{
return specified variable or default value if specified variable is null
}
