% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.get_item}
\alias{.get_item}
\title{Get item coresponding to given id and preprocess it's content}
\usage{
.get_item(id)
}
\arguments{
\item{id}{id of the item that should be retrieved}
}
\value{
list containing API respone
}
\description{
Get item coresponding to given id and preprocess it's content
}
