% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.send_request}
\alias{.send_request}
\title{Sends a request to the specified url and retrieves it's content.}
\usage{
.send_request(request_url)
}
\arguments{
\item{request_url}{url to be used}
}
\value{
parsed content of the response object
}
\description{
Sends a request to the specified url and retrieves it's content.
}
