% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_best_stories_ids}
\alias{get_best_stories_ids}
\title{Hacker News best stories ids}
\usage{
get_best_stories_ids()
}
\value{
list of best stories ids
}
\description{
Retrieves best stories ids using Hacker News API
}
\examples{
\donttest{
best_stories_ids <- get_best_stories_ids()
}

}
