% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_comments}
\alias{get_comments}
\title{Hacker News nested comments}
\usage{
get_comments(item)
}
\arguments{
\item{item}{item whose children (comments) will be retrieved}
}
\value{
dataframe of all comments under an item
}
\description{
Retrieves all comments under an item
using Hacker News API
}
\examples{
\donttest{
story <- get_item_by_id(21499889)
comments <- get_comments(story)
comments
}

}
