% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_comments_with_root}
\alias{get_comments_with_root}
\title{Hacker News nested comments with root comment}
\usage{
get_comments_with_root(item)
}
\arguments{
\item{item}{item whose children (comments) will be retrieved}
}
\value{
dataframe containing specified item and all comments under that item
}
\description{
Returns specified item and all comments under it. Comments are retrieved
using the Hacker News API
}
