% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hn_api_response.R
\name{get_content}
\alias{get_content}
\title{Retrieves the response content}
\usage{
get_content(x)
}
\arguments{
\item{x}{hn_api_response object to retrieve content from}
}
\value{
content of the given hn_api_response object
}
\description{
Retrieves the response content
}
