% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_item_by_id}
\alias{get_item_by_id}
\title{Retrieves the item corresponding to the specified id}
\usage{
get_item_by_id(id)
}
\arguments{
\item{id}{id of the item that should be retrieved}
}
\value{
item correspoing to the specified id
}
\description{
Retrieves the item corresponding to the specified id
}
