% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_items_by_ids}
\alias{get_items_by_ids}
\title{Retrieves items for given ids list}
\usage{
get_items_by_ids(ids)
}
\arguments{
\item{ids}{list of ids to be trimed}
}
\value{
list of hn_item objects
}
\description{
Retrieves items for given ids list
}
