% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_latest_ask_stories}
\alias{get_latest_ask_stories}
\title{Hacker News latest ask stories}
\usage{
get_latest_ask_stories()
}
\value{
list of latest ask HN stories ids
}
\description{
Retrives latest ask stories using Hacker News API
}
\examples{
ask_stories <- get_latest_ask_stories()
}
