% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_latest_ask_stories_ids}
\alias{get_latest_ask_stories_ids}
\title{Get latest ask HN stories}
\usage{
get_latest_ask_stories_ids()
}
\value{
list of latest ask HN stories ids
}
\description{
Get latest ask HN stories
}
