% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_latest_job_stories_ids}
\alias{get_latest_job_stories_ids}
\title{Hacker News latest job stories ids}
\usage{
get_latest_job_stories_ids()
}
\value{
list of latest job stories ids
}
\description{
Retrieves latest job stories ids using Hacker News API
}
\examples{
job_stories_ids <- get_latest_job_stories_ids()

}
