% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_latest_show_stories}
\alias{get_latest_show_stories}
\title{Hacker News latest show stories}
\usage{
get_latest_show_stories()
}
\value{
list of latest show HN stories
}
\description{
Retrives latest show stories using Hacker News API
}
\examples{
show_stories <- get_latest_show_stories()
}
