% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_max_item_id}
\alias{get_max_item_id}
\title{Hacker News item largest id}
\usage{
get_max_item_id()
}
\value{
current largest item id
}
\description{
Retrieves largest item id available via Hacker News API
}
\examples{
# retrieve the id of the latest item
get_max_item_id()

}
