% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_new_stories}
\alias{get_new_stories}
\title{Get new stories}
\usage{
get_new_stories(max_items = NULL)
}
\arguments{
\item{max_items}{Maximum number of items to retrive. If max_items = NULL, returns all available}
}
\value{
list of new stories
}
\description{
Get new stories
}
