% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_new_stories_ids}
\alias{get_new_stories_ids}
\title{Hacker News newest stories ids}
\usage{
get_new_stories_ids()
}
\value{
list of new stories ids
}
\description{
Retrives newest stories ids using Hacker News API
}
\examples{
new_stories_ids <- get_new_stories_ids()

}
