% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_top_stories}
\alias{get_top_stories}
\title{Get top stories}
\usage{
get_top_stories(max_items = NULL)
}
\arguments{
\item{max_items}{Maximum number of items to retrive. If max_items = NULL, returns all available}
}
\value{
List of top stories
}
\description{
Get top stories
}
