% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_updates}
\alias{get_updates}
\title{Get changed items and profiles}
\usage{
get_updates()
}
\value{
list of changed item ids and list of changed profile usernames
}
\description{
Get changed items and profiles
}
