% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hn_item.R
\name{is_hn_item}
\alias{is_hn_item}
\title{Checks whether the given object is of the class
hn_item}
\usage{
is_hn_item(x)
}
\arguments{
\item{x}{object to test if it is of the class hn_item}
}
\value{
TRUE if the object is of the class hn_item
}
\description{
Checks whether the given object is of the class
hn_item
}
