% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hn_user.R
\name{is_hn_user}
\alias{is_hn_user}
\title{Checks whether the given object is of the class
hn_user}
\usage{
is_hn_user(x)
}
\arguments{
\item{x}{object to test if it is of the class hn_user}
}
\value{
TRUE if the object is of the class hn_user
}
\description{
Checks whether the given object is of the class
hn_user
}
