% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_json}
\alias{parse_json}
\title{Parses a json response}
\usage{
parse_json(response)
}
\arguments{
\item{response}{response object to parse}
}
\value{
parsed content of the given response
}
\description{
Parses a json response
}
