% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.hackerlist}
\alias{print.hackerlist}
\title{Print for "hackerlist" type objects}
\usage{
\method{print}{hackerlist}(x, ...)
}
\arguments{
\item{x}{"hackerlist" type list}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print for "hackerlist" type objects
}
\examples{

hacker_news <- get_top_stories(max_items = 5)
print(hacker_news)
}
