% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.hackeruser}
\alias{print.hackeruser}
\title{Print for "hackeruser" type objects}
\usage{
\method{print}{hackeruser}(x, ...)
}
\arguments{
\item{x}{"hackerlist" type list}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print for "hackeruser" type objects
}
\examples{

hacker_user <- get_user_by_username('steveddaniels')
print(hacker_user)
}
