% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hn_item.R
\name{print.hn_item}
\alias{print.hn_item}
\title{Print for "hn_item" type objects}
\usage{
\method{print}{hn_item}(x, ...)
}
\arguments{
\item{x}{"hn_item" type list}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print for "hn_item" type objects
}
