% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hn_user.R
\name{print.hn_user}
\alias{print.hn_user}
\title{Print for "hn_user" type objects}
\usage{
\method{print}{hn_user}(x, ...)
}
\arguments{
\item{x}{"hn_user" type list}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print for "hn_user" type objects
}
