% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_ids_list}
\alias{trim_ids_list}
\title{Selects only limited number of ids}
\usage{
trim_ids_list(ids_list, max_items)
}
\arguments{
\item{ids_list}{list of ids to be trimmed}

\item{max_items}{maximum number of ids to retrieve}
}
\value{
trimmed list of ids
}
\description{
Selects only limited number of ids
}
