% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hn_api_response.R
\name{validate_hn_api_response}
\alias{validate_hn_api_response}
\title{Checks if the given response is not empty and that
it did not return an error http code.}
\usage{
validate_hn_api_response(hn_api_response)
}
\arguments{
\item{hn_api_response}{hn_api_response object to be checked}
}
\description{
Checks if the given response is not empty and that
it did not return an error http code.
}
