% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hn_item.R
\name{validate_hn_item}
\alias{validate_hn_item}
\title{Checks whether the given object is correctly defined
hn_item class}
\usage{
validate_hn_item(hn_item)
}
\arguments{
\item{hn_item}{object of the class hn_item}
}
\description{
Checks whether the given object is correctly defined
hn_item class
}
