% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hn_user.R
\name{validate_hn_user}
\alias{validate_hn_user}
\title{Checks whether the given object is correctly defined
hn_user class}
\usage{
validate_hn_user(hn_user)
}
\arguments{
\item{hn_user}{object of the class hn_user}
}
\description{
Checks whether the given object is correctly defined
hn_user class
}
