% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_item_by_id}
\alias{get_item_by_id}
\title{Get Hacker News by Id}
\usage{
get_item_by_id(id)
}
\arguments{
\item{id}{id of the item that should be retrieved}
}
\value{
item correspoing to the specified id
}
\description{
Retrieves the item
corresponding to specified id using Hacker News API
}
\examples{
\donttest{
h_id <- get_max_item_id()
hacker_article <- get_item_by_id(h_id)
}

}
