% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_latest_ask_stories_ids}
\alias{get_latest_ask_stories_ids}
\title{Hacker News latest ask stories ids}
\usage{
get_latest_ask_stories_ids()
}
\value{
list of latest ask Hacker News stories ids
}
\description{
Retrieves latest ask stories ids using Hacker News API
}
\examples{
# get ids of latest ask stories
ask_stories_ids <- get_latest_ask_stories_ids()
ask_stories_ids

}
