% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_latest_job_stories}
\alias{get_latest_job_stories}
\title{Hacker News latest job stories}
\usage{
get_latest_job_stories(max_items = NULL)
}
\arguments{
\item{max_items}{Maximum number of items to retrive.
If max_items = NULL, returns all available}
}
\value{
list of latest job stories
}
\description{
Retrives latest job stories using Hacker News API
}
\examples{
job_stories <- get_latest_job_stories()

}
