% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_latest_job_stories_ids}
\alias{get_latest_job_stories_ids}
\title{Get latest job stories}
\usage{
get_latest_job_stories_ids()
}
\value{
list of latest job stories ids
}
\description{
Get latest job stories
}
