% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_latest_show_stories}
\alias{get_latest_show_stories}
\title{Hacker News latest show stories}
\usage{
get_latest_show_stories(max_items = NULL)
}
\arguments{
\item{max_items}{Maximum number of items to retrive. If max_items = NULL, returns all available}
}
\value{
list of latest show HN stories
}
\description{
Retrives latest show stories using Hacker News API
}
\examples{
\dontrun{
show_stories <- get_latest_show_stories()
}
}
