% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_latest_show_stories}
\alias{get_latest_show_stories}
\title{Get latest show HN stories}
\usage{
get_latest_show_stories()
}
\value{
list of latest show HN stories ids
}
\description{
Get latest show HN stories
}
