% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_latest_show_stories_ids}
\alias{get_latest_show_stories_ids}
\title{Hacker News latest show stories ids}
\usage{
get_latest_show_stories_ids()
}
\value{
list of latest show Hacker News stories ids
}
\description{
Retrieves latest show stories ids using Hacker News API
}
\examples{
# get ids of latest show stories on Hacker News
show_stories_ids <- get_latest_show_stories_ids()
show_stories_ids

}
