% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_top_stories}
\alias{get_top_stories}
\title{Hacker News top stories}
\usage{
get_top_stories(max_items = NULL)
}
\arguments{
\item{max_items}{Maximum number of items to retrieve.
If max_items = NULL, returns all available stories}
}
\value{
List of top stories
}
\description{
Retrieves top stories using Hacker News API
}
\details{
Parallel api requests can be enabled by running
\code{future::plan(future::multiprocess)}
}
\examples{
# get the top story on Hacker News
top_story <- get_top_stories(max_items = 1)
top_story

\donttest{
# get all top stories on Hacker News
top_stories <- get_top_stories()
top_stories
}

}
