% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_top_stories_ids}
\alias{get_top_stories_ids}
\title{Hacker News top stories ids}
\usage{
get_top_stories_ids()
}
\value{
list of top stories ids
}
\description{
Retrieves top stories ids using Hacker News API
}
\examples{

# retrieve the ids of current top stories on Hacker News
top_stories_ids <- get_top_stories_ids()
top_stories_ids

}
