% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_updates}
\alias{get_updates}
\title{Hacker News updated profiles}
\usage{
get_updates()
}
\value{
list of changed item ids and list of changed profile usernames
}
\description{
Retrives changed item ids and changed profile usernames using Hacker News API
}
\examples{
hacker_updates <- get_updates()
print(hacker_updates)

}
