% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_user_by_username}
\alias{get_user_by_username}
\title{Get Hacker News user}
\usage{
get_user_by_username(username)
}
\arguments{
\item{username}{of the user that should be retrieved}
}
\value{
user corresponding to the specified username
}
\description{
Get user corresponding to the specified username
}
\examples{
# Retrieve information about the jl user
get_user_by_username("jl")

}
