% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.hackerlist}
\alias{print.hackerlist}
\title{Define print for "hackerlist" type objects}
\usage{
\method{print}{hackerlist}(x, ...)
}
\arguments{
\item{x}{"hackerlist" type list}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Define print for "hackerlist" type objects
}
