% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_datetime_origin}
\alias{to_datetime_origin}
\title{Converts numeric value into POSIXct datetime type}
\usage{
to_datetime_origin(x)
}
\arguments{
\item{x}{numeric value to be converted}
}
\value{
time object of class POSIXct
}
\description{
Converts numeric value into POSIXct datetime type
}
\details{
Hacker News API supplies time as a Unix Time,
which origin date is "1970-01-01". This value has to be
passed to as.POSIXct function in order to get
proper time representation
}
\seealso{
\code{\link[base]{as.POSIXct}}

\url{https://github.com/HackerNews/API}

\url{https://en.wikipedia.org/wiki/Unix_time}
}
